+++ lineman_api/config.py
@@ -1,5 +1,5 @@
 import os
 
-sql_db = 'sqlite:////tmp/test.db'
-HOST = os.environ.get('HOST', 'localhost')
+SQL_DB = os.environ.get("SQLALCHEMY_DATABASE_URI", 'sqlite:////tmp/test.db')
+HOST = os.environ.get('HOST')
 PORT = os.environ.get('PORT', 5002)
